+++ src/App.tsx
@@ -1,4 +1,40 @@
 import React from 'react';
-import { KeyboardInput } from './ClassComponents/KeyboardInput';
 
-export const App: React.FC = () => <KeyboardInput />;
+type State = {
+  pressedKey: string;
+};
+
+export class App extends React.Component<{}, State> {
+  state: State = {
+    pressedKey: '',
+  };
+
+  componentDidMount(): void {
+    window.addEventListener('keyup', this.handleKeyUp);
+  }
+
+  componentWillUnmount(): void {
+    window.removeEventListener('keyup', this.handleKeyUp);
+  }
+
+  handleKeyUp = (event: KeyboardEvent) => {
+    event.preventDefault();
+    this.setState({ pressedKey: event.key });
+  };
+
+  render() {
+    const { pressedKey } = this.state;
+
+    return (
+      <div className="App">
+        <p className="App__message">
+          {pressedKey !== '' ? (
+            <>The last pressed key is [{pressedKey}]</>
+          ) : (
+            <>Nothing was pressed yet</>
+          )}
+        </p>
+      </div>
+    );
+  }
+}
